Use modulo instead of mul+sub in __secs_to_tm
On x86 modulo is free when doing division, so this removes
a multiplication and at the cost of replacing a conditional
move with a conditional jump, but it still appears to be
faster.
(Similar architectures: nds32le)

ARM doesn't have modulo, instead an multiply-and-subtract
operation is done after the division, so the diffence
here is either none at all, or a move and a multiply-and-add
being replaced with a multiply-and-subtract.
(Similar architectures: or1k)

RISC-V on the other hand has a separate modulo
instruction and will perform a separate modulo instead of
an assignment, a multiplication, and an addition with
this change. GCC does change how the modulo operation is
realised depending on the optimisation level. I don't know
how this affects the performance, however a simple test on
x86 suggests that doing a modulo operations is actually
faster than assign–multiply–add.